requirements.txt
absl-py==1.3.0
brotlipy==0.7.0
cachetools==4.2.4
click==8.1.3
cloudpickle==1.3.0
conda==4.13.0
conda-build==3.21.9
Cython==0.29.32
docker-pycreds==0.4.0
fasteners==0.18
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
glfw==2.5.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gym==0.17.2
imageio==2.22.3
importlib-metadata==5.0.0
Jinja2==2.10.1
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-service==2.4.0
mujoco-py==2.0.2.11
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyglet==1.5.0
pynvml==11.4.1
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
smmap==5.0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
urllib3==1.26.12
wandb==0.13.5
Werkzeug==2.2.2
zipp==3.10.0